feat: Add new `generate` API and make `generateImage` deprecated
