+++ deps.ts
@@ -1,3 +1,2 @@
 export * as base64 from "https://deno.land/std@0.205.0/encoding/base64.ts";
-export * as io from "https://deno.land/std@0.205.0/io/mod.ts";
 export * from "https://deno.land/x/unknownutil@v3.10.0/mod.ts";
