+++ deps_deprecated.ts
@@ -1,0 +1,1 @@
+export * as io from "https://deno.land/std@0.205.0/io/mod.ts";
